refactor(onboarding): add GoalSuggestion interface and goal suggestions constants
