+++ apps/api/src/modules/onboarding/onboarding.service.ts
@@ -1,44 +1,8 @@
 import { Injectable } from '@nestjs/common';
 
-export interface GoalSuggestion {
-  label: string;
-  role_category: string;
-  description: string;
-  estimated_weeks: number;
-}
+import type { GoalSuggestion } from './onboarding.constants';
 
-const GOAL_SUGGESTIONS: GoalSuggestion[] = [
-  {
-    label: 'Backend Intern',
-    role_category: 'Backend',
-    description: 'Build production-grade REST APIs and understand server-side fundamentals.',
-    estimated_weeks: 16,
-  },
-  {
-    label: 'Frontend Developer',
-    role_category: 'Frontend',
-    description: 'Master React, CSS architecture, and modern web performance.',
-    estimated_weeks: 16,
-  },
-  {
-    label: 'iOS Developer',
-    role_category: 'Mobile',
-    description: 'Learn Swift, UIKit, SwiftUI, and publish apps to the App Store.',
-    estimated_weeks: 24,
-  },
-  {
-    label: 'DevOps Engineer',
-    role_category: 'DevOps',
-    description: 'Master CI/CD pipelines, Docker, Kubernetes, and cloud infrastructure.',
-    estimated_weeks: 20,
-  },
-  {
-    label: 'Data Analyst',
-    role_category: 'Data',
-    description: 'Learn SQL, Python, Pandas, and data visualization tools like Tableau.',
-    estimated_weeks: 16,
-  },
-];
+import { GOAL_SUGGESTIONS } from './onboarding.constants';
 
 @Injectable()
 export class OnboardingService {
@@ -47,7 +11,7 @@ export class OnboardingService {
       return GOAL_SUGGESTIONS;
     }
     return GOAL_SUGGESTIONS.filter(
-      (goal) => goal.role_category.toLowerCase() === roleCategory.toLowerCase(),
+      (goal) => goal.roleCategory.toLowerCase() === roleCategory.toLowerCase(),
     );
   }
 }
